chore(fahrenheit-celcius): use floating-point arithmetic

+++ fahrenheit-celcius-table.c
@@ -3,7 +3,7 @@
 /* print Fahrenheit-Celcius table
     for fahr = 0, 20, ..., 300 */
 int main() {
-  int fahr, celcius;
+  float fahr, celcius;
   int lower, upper, step;
 
   lower = 0;   /* lower limit of termperature table */
@@ -13,7 +13,7 @@ int main() {
   fahr = lower;
   while (fahr <= upper) {
     celcius = 5 * (fahr - 32) / 9;
-    printf("%3d %6d\n", fahr, celcius);
+    printf("%3.0f %6.1f\n", fahr, celcius);
     fahr = fahr + step;
   }
 }
